+++ rbx_types/src/lib.rs
@@ -17,6 +17,7 @@ mod lister;
 mod physical_properties;
 mod referent;
 mod shared_string;
+mod tags;
 mod variant;
 
 pub use attributes::*;
@@ -30,4 +31,5 @@ pub use faces::*;
 pub use physical_properties::*;
 pub use referent::*;
 pub use shared_string::*;
+pub use tags::*;
 pub use variant::*;