+++ rbx_xml/src/types/mod.rs
@@ -26,6 +26,7 @@ mod rect;
 mod referent;
 mod shared_string;
 mod strings;
+mod tags;
 mod udims;
 mod vectors;
 
@@ -49,6 +50,7 @@ use crate::{
 use self::{
     referent::{read_ref, write_ref},
     shared_string::{read_shared_string, write_shared_string},
+    tags::write_tags,
 };
 
 /// The `declare_rbx_types` macro generates the two big match statements that
@@ -105,6 +107,7 @@ macro_rules! declare_rbx_types {
 
                 Variant::Ref(value) => write_ref(writer, xml_property_name, *value, state),
                 Variant::SharedString(value) => write_shared_string(writer, xml_property_name, value, state),
+                Variant::Tags(value) => write_tags(writer, xml_property_name, value),
 
                 unknown => {
                     Err(writer.error(EncodeErrorKind::UnsupportedPropertyType(unknown.ty())))