+++ rbx_binary/CHANGELOG.md
@@ -1,6 +1,12 @@
 # rbx_binary Changelog
 
 ## Unreleased
+* Added support for reading `String` as `Tags` and writing `Tags` as `String`. ([#199][pr-199])
+
+[pr-199]: https://github.com/rojo-rbx/rbx-dom/pull/199
+
+## 0.6.2 (2021-07-19)
+* Upgraded to rbx\_dom\_weak 2.2.
 
 ## 0.6.1 (2021-07-02)
 * Upgraded to rbx\_dom\_weak 2.1.